+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+dns-update 0.2.5
+================================
+- BunnyDNS: Use subdomain as name of record instead of FQDN.
+- RFC2136: Chunk TXT records.
+
 dns-update 0.2.4
 ================================
 - Google Cloud DNS: Chunk TXT records into 255-character strings when updating records.